move the rest of datalake operations to use the operations macro
In the previous PR, these were missed.  Additionally, this splits out
the rename operation off of the path_put into it's own file.
